requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.0.1
scikit-learn==1.5.1
transformers==4.41.2
accelerate==0.31.0
pydantic==2.8.2
streamlit==1.36.0
requests==2.32.3
pandas==2.2.2
google-genai